requirements/base.txt
Django==2.1.3
django-rest-auth==0.9.3
djangorestframework==3.9.0
dnspython==1.15.0
pytz==2018.7
six==1.11.0
